refactor : Car, Cars, CarsFactory의 패키지 구조 및 접근제어자 수정
Car와 Cars의 생성은 CarsFactory를 통해서만 할 수 있도록, 패키지 구조를 변경하고 접근제어자를 변경
