+++ src/main/java/racingcar/Car/CarsFactory.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.Car;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -11,8 +11,8 @@ public class CarsFactory {
     private static final String ERROR_ID_MAX_LENGTH_MSG = "[ERROR] 사용자의 아이디는 5글자 이하여야 합니다.";
     private static final String ERROR_ID_DUPLICATE_MSG = "[ERROR] 중복된 사용자 이름입니다.";
     private static final String ERROR_MAX_SIZE_OF_PEOPLE_MSG = "[ERROR] 인원숫자를 확인해주십시오. (2명 이상 5명 이하)";
-    public static final int MIN_PEOPLE_SIZE = 2;
-    public static final int MAX_PEOPLE_SIZE = 5;
+    private static final int MIN_PEOPLE_SIZE = 2;
+    private static final int MAX_PEOPLE_SIZE = 5;
 
     public static Cars createCars(String[] carNames) {
         validateCarNamesSize(carNames);
