+++ src/main/java/racingcar/Racing.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.Car.Cars;
+
 import java.util.List;
 
 public class Racing {
